+++ packages/service-env/.gitignore
@@ -1,2 +1,3 @@
 /*.js
+/lib
 yarn.lock
