+++ packages/service-env/package.json
@@ -4,7 +4,7 @@
   "version": "0.0.39-alpha.0",
   "author": "Marco Pegoraro <marco.pegoraro@gmail.com",
   "bugs": {
-    "url": "https://github.com/forrestjs/core/issues"
+    "url": "https://github.com/forrestjs/forrestjs/issues"
   },
   "dependencies": {
     "@babel/runtime": "^7.0.0",
@@ -17,18 +17,19 @@
     "@forrestjs/babel-preset-package": "^0.0.39-alpha.0",
     "cross-env": "^5.1.4"
   },
-  "homepage": "https://github.com/forrestjs/core/tree/master/packages/service-env#readme",
+  "homepage": "https://github.com/forrestjs/forrestjs/tree/master/packages/service-env#readme",
   "keywords": [
     "forrestjs",
-    "forrestjs-feature",
+    "service",
     "env",
-    "dotenv"
+    "dotenv",
+    "environment"
   ],
   "license": "MIT",
   "publishConfig": {
     "access": "public"
   },
-  "repository": "https://github.com/forrestjs/core/tree/master/packages/service-env",
+  "repository": "https://github.com/forrestjs/forrestjs/tree/master/packages/service-env",
   "scripts": {
     "build": "babel src --out-dir . --ignore **/__tests__",
     "prepare": "cross-env NODE_ENV=production npm run build",
